feat(DeviceConnector/Device): Proper monitoring & device handling Part 1
