+++ packages/device-connector/src/deviceConnector.ts
@@ -1,10 +1,12 @@
 import { connectCli, registerDevices } from './arduino-cli/service'
 import { RPCClient } from './arduino-cli/client'
-import { deployBinary } from './devices/service'
+import { deployBinary } from './devices/deployment'
 import { openStream } from './deployment-server/connection'
 import logger from './util/logger'
 import { Signals } from 'close-with-grace'
 import { Duplex } from 'stream'
+import { DeployServRequest } from './deployment-server/service'
+import { monitorDevice } from './devices/monitoring'
 
 export interface DeviceConnector {
   client: RPCClient
@@ -18,12 +20,14 @@ export const createConnector = async (): Promise<DeviceConnector> => {
   await registerDevices(client)
 
   socket.on('data', (message) => {
-    const servReq = JSON.parse(message)
+    const servReq = JSON.parse(message) as DeployServRequest
     const type = servReq.type
     const data = servReq.data
 
     if (type === 'deploy') {
-      deployBinary(data, client).catch((e) => {
+      deployBinary(data, client).then(async (device) => {
+        await monitorDevice(client, device)
+      }).catch((e) => {
         socket.write(e.message)
         logger.error(e.message)
       })
